+++ backend/app/settings.py
@@ -203,6 +203,7 @@ CKEDITOR_CONFIGS = {
     'default': {
         'toolbar': 'full',
         'width': '100%',
+        'allowedContent': True
     },
 }
 
@@ -260,3 +261,6 @@ if DEBUG_TOOLBAR:
     ]
 
 TEST_COVERAGE_RATE = 50
+SENTRY_CHECK_METHOD = 'warning'
+
+DEFAULT_AUTO_FIELD = 'django.db.models.AutoField'