+++ backend/garpix_notify/admin/template.py
@@ -1,6 +1,7 @@
-from django.contrib import admin
+from django.contrib import admin, messages
+
+from ..models import SystemNotify, Notify
 from ..models.template import NotifyTemplate
-from ..models.notify import Notify
 from django.http import HttpResponseRedirect
 
 
@@ -11,6 +12,7 @@ class NotifyTemplateAdmin(admin.ModelAdmin):
         'title',
         'is_active',
         'subject',
+        'is_delete_after',
         'get_context_description',
         'text',
         'html',
@@ -33,28 +35,47 @@ class NotifyTemplateAdmin(admin.ModelAdmin):
     list_filter = ('type', 'category', 'event', 'is_active')
     actions = ['create_mailing', ]
     filter_horizontal = ('user_lists',)
+    raw_id_fields = ('user',)
 
     def create_mailing(self, request, queryset):
         count = Notify.send(event=None, context={}, notify_templates=queryset)
         self.message_user(request, 'Рассылка создана, кол-во сообщений: {}'.format(count))
+
     create_mailing.short_description = "Сделать рассылку"
 
     def response_change(self, request, obj):
-        from ..models.notify import Notify
+        context = obj.get_test_data()
+        template = obj
+        user = obj.user if obj.user else request.user
         if obj.user_lists and "_send_now" in request.POST:
-
-            context = obj.get_test_data()
-            template = obj
             instance = Notify.objects.create(
                 subject=obj.render_subject(template.subject),
                 text=obj.render_text(context),
                 html=obj.render_html(context),
-                user=obj.user,
+                user=user,
                 email=obj.email,
                 type=obj.type,
                 category=obj.category,
             )
-            instance.save()
-            instance._start_send()
+            instance.start_send()
+            self.message_user(request, 'Тестовое уведомление отправлено', level=messages.SUCCESS)
             return HttpResponseRedirect(".")
+        elif "_send_now_system" in request.POST:
+            instance = SystemNotify.objects.create(
+                title=template.subject if template.subject or template.subject != '' else template.title,
+                event=template.event,
+                user=user,
+                data_json=context,
+                room_name=f'room_{user.pk}'
+            )
+            instance.send_notification()
+            self.message_user(request, 'Тестовое уведомление отправлено', level=messages.SUCCESS)
+            return HttpResponseRedirect(".")
+
         return super().response_change(request, obj)
+
+    def get_changelist(self, request, **kwargs):
+        events_message = NotifyTemplate.get_blank_events_message()
+        if events_message:
+            self.message_user(request, events_message, level=messages.WARNING)
+        return super().get_changelist(request, **kwargs)