+++ backend/garpix_notify/models/log.py
@@ -2,7 +2,8 @@ from django.db import models
 
 
 class NotifyErrorLog(models.Model):
-    notify = models.ForeignKey('garpix_notify.Notify', on_delete=models.CASCADE, related_name='logs', verbose_name='Notify')
+    notify = models.ForeignKey('garpix_notify.Notify', on_delete=models.CASCADE, related_name='logs',
+                               verbose_name='Notify')
     error = models.TextField(verbose_name='Ошибка', blank=True, null=True, default='')
     created_at = models.DateTimeField(auto_now_add=True, verbose_name='Дата создания')